python/requirements.txt
appdirs==1.4.4
appdynamics-lambda-tracer==20.7.383
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Faker==4.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pathspec==0.8.0
PyMySQL==0.10.0
python-dateutil==2.8.1
regex==2020.7.14
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
